requirements.txt
albumentations==1.1.0
altair==4.2.0
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bleach==5.0.0
blinker==1.4
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
efficientnet-pytorch==0.7.1
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
imageio==2.19.1
importlib-metadata==4.11.3
importlib-resources==5.7.1
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.5.1
jupyter-client==7.3.1
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8
notebook==6.4.11
numpy==1.22.3
opencv-python-headless==4.5.5.64
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==8.0.0
pycparser==2.21
pydeck==0.7.1
Pygments==2.12.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==22.3.0
qudida==0.0.4
requests==2.27.1
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
semver==2.13.0
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
soupsieve==2.3.2.post1
stack-data==0.2.0
streamlit==1.9.0
terminado==0.13.3
threadpoolctl==3.1.0
tifffile==2022.5.4
tinycss2==1.1.1
toml==0.10.2
toolz==0.11.2
torch==1.11.0
torchvision==0.12.0
tornado==6.1
traitlets==5.2.0
typing-extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
validators==0.19.0
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.6.0
zipp==3.8.0
